feat: add model load balancing configs and config entry modal
